+++ src/main/java/wooteco/subway/path/event/handler/LineUpdateEventHandler.java
@@ -3,7 +3,7 @@ package wooteco.subway.path.event.handler;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 import wooteco.subway.line.event.LineUpdatedEvent;
-import wooteco.subway.path.domin.PathRepository;
+import wooteco.subway.path.domain.PathRepository;
 
 @Component
 public class LineUpdateEventHandler {
